Move quill config stuff to separate file
